Repoisitory 생성  - WiseSayingReposity
